backend/requirements.txt
Flask==2.0.1
mysql-connector==2.2.9
faker